+++ src/irvine/math/series/Integral.java
@@ -11,21 +11,21 @@ class Integral<E> implements Series<E> {
 
   private final Field<E> mElementField;
   private final Series<E> mS;
-  private final int mBound;
+  private final long mBound;
 
   Integral(final Field<E> elementField, final Series<E> s) {
     mElementField = elementField;
     mS = s;
-    mBound = mS.bound() == Integer.MAX_VALUE ? Integer.MAX_VALUE : mS.bound() + 1;
+    mBound = mS.bound() == Long.MAX_VALUE ? Long.MAX_VALUE : mS.bound() + 1;
   }
 
   @Override
-  public E coeff(final int n) {
+  public E coeff(final long n) {
     return n == 0 ? mElementField.zero() : mElementField.divide(mS.coeff(n - 1), mElementField.coerce(n));
   }
 
   @Override
-  public int bound() {
+  public long bound() {
     return mBound;
   }
 }
