+++ src/irvine/math/series/Monomial.java
@@ -9,21 +9,21 @@ class Monomial<E> implements Series<E> {
 
   private final E mZero;
   private final E mA;
-  private final int mN;
+  private final long mN;
 
-  Monomial(final E zero, final E a, final int n) {
+  Monomial(final E zero, final E a, final long n) {
     mZero = zero;
     mA = a;
     mN = n;
   }
 
   @Override
-  public E coeff(final int n) {
+  public E coeff(final long n) {
     return n == mN ? mA : mZero;
   }
 
   @Override
-  public int bound() {
+  public long bound() {
     return mN;
   }
 }
