+++ src/irvine/math/series/SubstitutePolynomial.java
@@ -15,7 +15,7 @@ class SubstitutePolynomial<E> implements Series<E> {
   private final Field<E> mElementField;
   private final Polynomial<E> mS;
   private final Series<E> mT;
-  private final int mBound;
+  private final long mBound;
 
   SubstitutePolynomial(final SeriesRing<E> ring, final Polynomial<E> s, final Series<E> t) {
     mRing = ring;
@@ -23,16 +23,16 @@ class SubstitutePolynomial<E> implements Series<E> {
     mS = s;
     mT = t;
     final Z b = Z.valueOf(mT.bound()).multiply(mS.degree());
-    mBound = b.bitLength() < Integer.SIZE ? b.intValue() : Integer.MAX_VALUE;
+    mBound = b.bitLength() < Long.SIZE ? b.longValue() : Long.MAX_VALUE;
   }
 
   @Override
-  public E coeff(final int n) {
+  public E coeff(final long n) {
     // [x^n] s(t(x)) = Sum_{k=0..n} s(k) * [x^n] t(x)^k
     E sum = mElementField.zero();
     if (n <= mBound) {
       Series<E> tk = mRing.one();
-      for (int k = 0; k <= mS.degree(); ++k) {
+      for (long k = 0; k <= mS.degree(); ++k) {
         final E sk = mS.coeff(k);
         if (!mElementField.isZero(sk)) {
           sum = mElementField.add(sum, mElementField.multiply(sk, tk.coeff(n)));
@@ -44,7 +44,7 @@ class SubstitutePolynomial<E> implements Series<E> {
   }
 
   @Override
-  public int bound() {
+  public long bound() {
     return mBound;
   }
 }
